Changed temp file deletion in xv, to leave files that have been generated in less than 1 minute ago

+++ pandas.py
@@ -5,6 +5,7 @@ Tools for Pandas
 import os
 import tempfile
 import glob
+import time
 from datetime import datetime as dt
 
 import pandas as pd
@@ -77,7 +78,9 @@ def xview(df, index=True, label=''):
 
     try:
         for file in oldfiles:
-            os.remove(os.path.abspath(file))
+            time_alive = now = time.time() - os.path.getctime(file)
+            if time_alive > 60:
+                os.remove(os.path.abspath(file))
     except:
         pass
